WIP on Teensy detection.
This is a somewhat failed attempt caused by a problem in the Teensy
installer: The Teensyduino installer requires you to type in the path
to the Arduino software so that it can patch it, and does not provide
a way to supply that path via a command line option. Thus we cannot
automate it yet. May have to give up and stop using Teensyduino if we
cannot get a find a workaround for that problem.